scrapers/requirements.txt
selenium==4.18.1
beautifulsoup4==4.12.3
pandas==2.2.1
requests==2.31.0
webdriver-manager==4.0.1
fake-useragent==1.4.0
python-dotenv==1.0.1
tqdm==4.66.2
pillow==10.2.0
schedule==1.2.1
pymysql==1.1.0
sqlalchemy==2.0.28
colorama==0.4.6 